adds fixes for memory usage for gc
